# api/swagger.yml
openapi: 3.1.0
info:
  title: Receipt Wrangler API.
  termsOfService:
    there are no TOS at this moment, use at your own risk we take no
    responsibility
  license:
    name: GPLV3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 5.0.0
servers:
  - url: /api
security:
  - bearerAuth: [ ]
  - apiKeyAuth: [ ]
paths:
...
  /systemTask/getPagedSystemTasks:
    post:
      tags:
        - SystemTask
      summary: Gets paged system tasks
      description: This will return paged system tasks
      operationId: getPagedSystemTasks
      requestBody:
        description: Paging and sorting data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetSystemTaskCommand"
      responses:
        200:
          description: Paged system tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedData"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
  /systemTask/getPagedActivities:
    post:
      tags:
        - SystemTask
      summary: Gets paged activities
      description: This will return paged activities for a list of groups
      operationId: getPagedActivities
      requestBody:
        description: Paging and sorting data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagedActivityRequestCommand"
      responses:
        200:
          description: Paged activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedData"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
  /systemTask/rerunActivity/{id}:
    post:
      tags:
        - SystemTask
      summary: Attempts to rerun activity
      description: This will rerun a failed activity
      operationId: rerunActivity
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Id of activity to restart
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API Key in format v1.xxx.xxx.xxx
  responses:
    Ok:
      description: Request was successfully processed
    BadRequest:
      description: The request was malformed
...
    Forbidden:
      description: The request was not allowed
...
    Internal:
      description: There was an error processing the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalErrorResponse'
  schemas:
...
    ReceiptStatus:
      description: Status of a receipt
      type: string
      enum:
        - "OPEN"
        - "NEEDS_ATTENTION"
        - "RESOLVED"
        - "DRAFT"
        - ""
    QuickScanDefaultPaidByType:
      description: How the default paid-by is resolved for optional quick-scan paid-by
      type: string
      enum:
        - "UPLOADER"
        - "USER"
        - ""
...
    ItemStatus:
      type: string
      enum:
        - "OPEN"
        - "RESOLVED"
        - "DRAFT"
    GroupStatus:
      type: string
      enum:
        - "ACTIVE"
        - "ARCHIVED"
...
    SortDirection:
      type: string
      enum:
        - asc
        - desc
        - ""
...
    AiType:
      type: string
      enum:
        - "OPEN_AI_CUSTOM"
        - "OPEN_AI"
        - "GEMINI"
        - "OLLAMA"
    OcrEngine:
      type: string
      enum:
        - "TESSERACT"
        - "EASY_OCR"
    SystemTaskStatus:
      type: string
      enum:
        - "SUCCEEDED"
        - "FAILED"
    SystemTaskType:
      type: string
      enum:
        - "OCR_PROCESSING"
        - "CHAT_COMPLETION"
        - "MAGIC_FILL"
        - "QUICK_SCAN"
        - "EMAIL_READ"
        - "EMAIL_UPLOAD"
        - "SYSTEM_EMAIL_CONNECTIVITY_CHECK"
        - "RECEIPT_PROCESSING_SETTINGS_CONNECTIVITY_CHECK"
        - "RECEIPT_UPLOADED"
        - "RECEIPT_UPDATED"
        - "PROMPT_GENERATED"
        - "API_KEY_DELETED"
        - "HTML_TO_PDF"
    AssociatedEntityType:
      type: string
      enum:
        - "NOOP_ENTITY_TYPE"
        - "RECEIPT"
        - "SYSTEM_EMAIL"
        - "RECEIPT_PROCESSING_SETTINGS"
        - "PROMPT"
        - "API_KEY"
...
    CustomFieldType:
      type: string
      enum:
        - "TEXT"
        - "DATE"
        - "SELECT"
        - "CURRENCY"
        - "BOOLEAN"
    Activity:
      type: object
      required:
        - id
        - type
        - status
        - startedAt
        - endedAt
        - childSystemTasks
      properties:
        id:
          type: integer
        type:
          $ref: "#/components/schemas/SystemTaskType"
        status:
          $ref: "#/components/schemas/SystemTaskStatus"
        startedAt:
          type: string
        endedAt:
          type: string
        ranByUserId:
          type: integer
          format: uint64
        receiptId:
          type: integer
          format: uint64
        groupId:
          type: integer
          format: uint64
        canBeRestarted:
          type: boolean
    BaseModel:
      type: object
      required:
        - id
        - createdAt
      properties:
        id:
          type: integer
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
          default: 0
        createdByString:
          type: string
          description: Created by entity's name
          default: ""
        updatedAt:
          type: string
          default: ""
    Category:
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the category
        description:
          type: string
          description: Description of the category
        updatedAt:
          type: string
          x-go-name: UpdatedAt
      description: Category to relate receipts to
...
    Comment:
      required:
        - comment
        - receiptId
        - userId
        - id
      type: object
      properties:
        additionalInfo:
          type: string
          description: Additional information about the comment
        comment:
          type: string
          description: Comment itself
        commentId:
          type: integer
          description: Comment foreign key used for repleis
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        receiptId:
          type: integer
          description: Receipt foreign key
        updatedAt:
          type: string
        userId:
          type: integer
          description: User foreign key
          format: uint64
      description: User comment left on receipts
    FileData:
      required:
        - receiptId
        - id
      type: object
      properties:
        createdAt:
          type: string

          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
          x-go-name: CreatedBy
        fileType:
          type: string
          description: MIME file type
          x-go-name: FileType
        id:
          type: integer
          format: uint64
          x-go-name: ID
        imageData:
          type: array
          description: Image data
          items:
            type: integer
            format: uint8
          x-go-name: ImageData
        name:
          type: string
          description: File name
          x-go-name: Name
        receiptId:
          type: integer
          description: Receipt foreign key
          format: uint64
          x-go-name: ReceiptId
        size:
          type: integer
          description: File size
          format: uint64
          x-go-name: Size
        updatedAt:
          type: string

          x-go-name: UpdatedAt
      description: File data for images on a receipt
      x-go-package: receipt-wrangler/api/internal/models
    Group:
      required:
        - name
        - status
        - id
        - groupMembers
        - isAllGroup
        - groupReceiptSettings
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        groupSettings:
          $ref: "#/components/schemas/GroupSettings"
        groupReceiptSettings:
          $ref: "#/components/schemas/GroupReceiptSettings"
        groupMembers:
          type: array
          description: Members of the group
          items:
            $ref: "#/components/schemas/GroupMember"
        id:
          type: integer
          format: uint64
        isDefault:
          type: boolean
          description: Is default group (not used yet)
        name:
          type: string
          description: Name of the group
        isAllGroup:
          type: boolean
          description: Is all group for user
        status:
          type: string
          $ref: "#/components/schemas/GroupStatus"
        updatedAt:
          type: string
      description: Group in the system
...
    GroupMember:
      required:
        - groupId
        - userId
      type: object
      properties:
        createdAt:
          type: string
        groupId:
          type: integer
          description: Group compound primary key
          format: uint64
        groupRoleId:
          type: integer
          format: uint64
          nullable: true
          description: Id of the modern group role assigned to the member
        updatedAt:
          type: string
          x-go-name: UpdatedAt
        userId:
          type: integer
          description: User compound primary key
          format: uint64
          x-go-name: UserID
      description: Group member
...
    Item:
      required:
        - amount
        - name
        - receiptId
        - status
      type: object
      properties:
        IsTaxed:
          type: boolean
          description: Is taxed (not used)
        amount:
          type: string
          description: Amount the item costs
        chargedToUserId:
          type: integer
          description: User foreign key
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Item name
        receiptId:
          type: integer
          description: Receipt foreign key
          format: uint64
        status:
          type: string
          $ref: "#/components/schemas/ItemStatus"
        linkedItems:
          type: array
          description: Items linked to this item (for sharing)
          items:
            $ref: "#/components/schemas/Item"
        categories:
          type: array
          description: Categories associated to the item
          items:
            $ref: "#/components/schemas/Category"
        tags:
          type: array
          description: Tags associated to the item
          items:
            $ref: "#/components/schemas/Tag"
        updatedAt:
          type: string
      description: Itemized item on a receipt
      x-go-package: receipt-wrangler/api/internal/models
...
    Receipt:
      required:
        - amount
        - date
        - groupId
        - name
        - paidByUserId
        - status
        - customFields
        - categories
        - tags
        - receiptItems
        - comments
        - id
      type: object
      properties:
        amount:
          type: string
          description: Receipt total amount
        categories:
          type: array
          description: Categories associated to receipt
          items:
            $ref: "#/components/schemas/Category"
        comments:
          type: array
          description: Comments associated to receipt
          items:
            $ref: "#/components/schemas/Comment"
        customFields:
          type: array
          description: Custom fields associated to receipt
          items:
            $ref: "#/components/schemas/CustomFieldValue"
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        date:
          type: string
          description: Receipt date
        groupId:
          type: integer
          description: Group foreign key
          format: uint64
        id:
          type: integer
          format: uint64
        imageFiles:
          type: array
          description: Files associated to receipt
          items:
            $ref: "#/components/schemas/FileData"
        name:
          type: string
          description: Receipt name
        paidByUserId:
          type: integer
          description: User paid foreign key
          format: uint64
        receiptItems:
          type: array
          description: Items associated to receipt
          items:
            $ref: "#/components/schemas/Item"
        resolvedDate:
          type: string
          description: Date resolved
        status:
          $ref: "#/components/schemas/ReceiptStatus"
        tags:
          type: array
          description: Tags associated to receipt
          items:
            $ref: "#/components/schemas/Tag"
        updatedAt:
          type: string
        createdByString:
          type: string
          description: Created by string, which is anything that is not a user
      description: Receipt
    Tag:
      required:
        - name
      type: object
      properties:
        createdAt:
          type: string
          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Tag name
        description:
          type: string
          description: Tag description
        updatedAt:
          type: string
      description: Tag to relate receipts to
...
    TagView:
      required:
        - id
        - name
        - numberOfReceipts
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the tag
        description:
          type: string
          description: Description of the tag
        updatedAt:
          type: string
        numberOfReceipts:
          type: integer
          description: Number of receipts associated with this tag
      description: Tag to relate receipts to
...
    PagedRequestCommand:
      required:
        - page
        - pageSize
      type: object
      properties:
        page:
          type: integer
          description: Page number
        pageSize:
          type: integer
          description: Number of records per page
        orderBy:
          type: string
          description: field to order on
        sortDirection:
          type: string
          $ref: "#/components/schemas/SortDirection"
...
    PagedActivityRequestCommand:
      allOf:
        - $ref: "#/components/schemas/PagedRequestCommand"
        - type: object
          properties:
            groupIds:
              type: array
              items:
                type: integer
...
    PagedData:
      required:
        - data
        - totalCount
      type: object
      properties:
        data:
          type: array
          items:
            anyOf:
              - $ref: "#/components/schemas/Receipt"
              - $ref: "#/components/schemas/Category"
              - $ref: "#/components/schemas/Tag"
              - $ref: "#/components/schemas/Prompt"
              - $ref: "#/components/schemas/Group"
              - $ref: "#/components/schemas/TagView"
              - $ref: "#/components/schemas/SystemTask"
              - $ref: "#/components/schemas/ReceiptProcessingSettings"
              - $ref: "#/components/schemas/SystemEmail"
              - $ref: "#/components/schemas/Activity"
              - $ref: "#/components/schemas/CustomField"
              - $ref: "#/components/schemas/ReportTemplate"
        totalCount:
          type: integer
...
    GroupReceiptSettings:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - groupId
          properties:
            groupId:
              type: integer
              description: Group foreign key
            hideImages:
              type: boolean
              description: Hide receipt images
            hideReceiptCategories:
              type: boolean
              description: Hide receipt categories
            hideReceiptTags:
              type: boolean
              description: Hide receipt tags
            hideItemCategories:
              type: boolean
              description: Hide receipt item categories
            hideItemTags:
              type: boolean
              description: Hide receipt item tags
            hideComments:
              type: boolean
              description: Hide receipt comments
            hideShareCategories:
              type: boolean
              description: Hide share categories
            hideShareTags:
              type: boolean
              description: Hide share tags
            quickScanPaidByEnabled:
              type: boolean
              description: Show the paid by field in quick scan
            quickScanPaidByRequired:
              type: boolean
              description: Require the paid by field in quick scan
            quickScanDefaultPaidByType:
              $ref: "#/components/schemas/QuickScanDefaultPaidByType"
            quickScanDefaultPaidById:
              type: integer
              nullable: true
              description: Default paid by user id when paid by is optional and type is USER
            quickScanStatusEnabled:
              type: boolean
              description: Show the status field in quick scan
            quickScanStatusRequired:
              type: boolean
              description: Require the status field in quick scan
            quickScanDefaultStatus:
              $ref: "#/components/schemas/ReceiptStatus"
            quickScanCategoriesEnabled:
              type: boolean
              description: Show the categories field in quick scan
            quickScanCategoriesRequired:
              type: boolean
              description: Require the categories field in quick scan
            quickScanTagsEnabled:
              type: boolean
              description: Show the tags field in quick scan
            quickScanTagsRequired:
              type: boolean
              description: Require the tags field in quick scan
...
    GroupSettings:
      type: object
      required:
        - id
        - groupId
      properties:
        id:
          type: integer
          description: Group settings id
        groupId:
          type: integer
          description: Group foreign key
        emailIntegrationEnabled:
          type: boolean
          description: Whether email integration is enabled
        emailBodyProcessingEnabled:
          type: boolean
          description: Whether email body text processing is enabled (opt-in, default false)
        systemEmailId:
          type: integer
          description: System email foreign key
        systemEmail:
          $ref: "#/components/schemas/SystemEmail"
        emailToRead:
          type: string
          description: Email to read
        subjectLineRegexes:
          type: array
          description: Subject line regexes
          items:
            $ref: "#/components/schemas/SubjectLineRegex"
        emailWhiteList:
          type: array
          description: Email white list
          items:
            $ref: "#/components/schemas/GroupSettingsWhiteListEmail"
        emailDefaultReceiptStatus:
          type: string
          $ref: "#/components/schemas/ReceiptStatus"
          description: Default receipt status
        emailDefaultReceiptPaidById:
          type: integer
          description: User foreign key
        prompt:
          $ref: "#/components/schemas/Prompt"
        promptId:
          type: integer
          description: Prompt foreign key
        fallbackPrompt:
          $ref: "#/components/schemas/Prompt"
        fallbackPromptId:
          type: integer
          description: Fallback prompt foreign key
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        updatedAt:
          type: string
...
    SystemEmail:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            host:
              type: string
              description: IMAP host
            port:
              type: string
              description: IMAP port
            username:
              type: string
              description: IMAP username
            password:
              type: string
              description: IMAP password
            useStartTLS:
              type: boolean
              description: Whether to use STARTTLS
...
    SystemTask:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            type:
              $ref: "#/components/schemas/SystemTaskType"
            status:
              $ref: "#/components/schemas/SystemTaskStatus"
            startedAt:
              type: string
            endedAt:
              type: string
            associatedEntityId:
              type: integer
            associatedEntityType:
              $ref: "#/components/schemas/AssociatedEntityType"
            ranByUserId:
              type: integer
              format: uint64
            receiptId:
              type: integer
              format: uint64
            groupId:
              type: integer
              format: uint64
            resultDescription:
              type: string
            apiKeyId:
              type: string
              nullable: true
            childSystemTasks:
              type: array
              items:
                $ref: "#/components/schemas/SystemTask"
    GetSystemTaskCommand:
      allOf:
        - $ref: "#/components/schemas/PagedRequestCommand"
        - type: object
      properties:
        associatedEntityId:
          type: integer
          description: Associated entity id
        associatedEntityType:
          $ref: "#/components/schemas/AssociatedEntityType"
    ReceiptProcessingSettings:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            name:
              type: string
              description: Name of the settings
            description:
              type: string
              description: Description of the settings
            aiType:
              $ref: "#/components/schemas/AiType"
            url:
              type: string
              description: URL for custom endpoints
            key:
              type: string
              description: Key for endpoints that require authentication
            model:
              type: string
              description: LLM model
            isVisionModel:
              type: boolean
              description: Is vision model
            enforceJsonResponseFormat:
              type: boolean
              description: Enforce JSON response format on the LLM provider. Disable if the provider does not support this flag.
            ocrEngine:
              $ref: "#/components/schemas/OcrEngine"
            prompt:
              $ref: "#/components/schemas/Prompt"
            promptId:
              type: integer
              description: Prompt foreign key
...
    Prompt:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - name
            - prompt
          properties:
            name:
              type: string
              description: Prompt name
            description:
              type: string
              description: Prompt description
            prompt:
              type: string
              description: Prompt text
...
    CustomFieldOption:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - customFieldId
          properties:
            value:
              type: string
              description: Custom Field Option value
            customFieldId:
              type: integer
              description: Custom Field Id
...
    CustomField:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - name
            - type
          properties:
            name:
              type: string
              description: Custom Field name
            type:
              $ref: "#/components/schemas/CustomFieldType"
            description:
              type: string
              description: Custom Field description
            options:
              type: array
              items:
                $ref: "#/components/schemas/CustomFieldOption"
...
    CustomFieldValue:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - receiptId
            - customFieldId
          properties:
            receiptId:
              type: integer
              description: Receipt Id
            customFieldId:
              type: integer
              description: Custom Field ID
            customField:
              $ref: "#/components/schemas/CustomField"
            stringValue:
              type: string
              description: Custom Field String Value
            dateValue:
              type: string
              description: Custom Field Date Value
            selectValue:
              type: integer
              description: Custom Field Select Value
            currencyValue:
              type: string
              description: Custom Field Currency Value
            booleanValue:
              type: boolean
              description: Custom Field Boolean Value
...
    InternalErrorResponse:
      type: object
      required:
        - errorMsg
      properties:
        errorMsg:
          type: string
          description: Error message
...
    ReportTemplate:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - name
            - configuration
            - configurationVersion
          properties:
            name:
              type: string
              description: The template name (mirrors the saved report's name).
            configuration:
              $ref: "#/components/schemas/ReportRequestCommand"
            configurationVersion:
              type: integer
              description: Schema version the stored configuration was written under.
            allowedActions:
              type: array
              description: >-
                The actions the requesting user may perform on this template
                (read, generate, update, delete, duplicate), resolved per user and
                populated only on the list response. Drives the row action buttons.
              items:
                type: string
...